# Fix theme-dependend markup extensions not knowing current theme context