--- a/src/Markup/Avalonia.Markup.Xaml/Styling/ResourceInclude.cs
+++ b/src/Markup/Avalonia.Markup.Xaml/Styling/ResourceInclude.cs
@@ -13,7 +13,7 @@ namespace Avalonia.Markup.Xaml.Styling
     /// When used in runtime, this type might be unsafe with trimming and AOT.
     /// </remarks>
     [RequiresUnreferencedCode(TrimmingMessages.StyleResourceIncludeRequiresUnreferenceCodeMessage)]
-    public class ResourceInclude : IResourceProvider
+    public class ResourceInclude : IResourceProvider, IThemeVariantProvider
     {
         private readonly IServiceProvider? _serviceProvider;
         private readonly Uri? _baseUri;
@@ -65,6 +65,8 @@ namespace Avalonia.Markup.Xaml.Styling
         /// </summary>
         public Uri? Source { get; set; }
 
+        ThemeVariant? IThemeVariantProvider.Key { get; set; }
+        
         bool IResourceNode.HasResources => Loaded.HasResources;
 
         public event EventHandler? OwnerChanged