## What does the pull request do?

Makes ThemeDictionaries lookup closer to the UWP behavior.
StaticResource (as well as floating DynamicResource) defined in the ThemeDictionary should use local theme variant instead of global application one (nor Default).

## How was the solution implemented (if it's not obvious)?

I was thinking about adding a new IServiceProvider service, which would require adding support to the XamlX. I wanted to avoid that if possible as this service would be very Avalonia specific.
In this PR approach with interface and simple compiler transformer works well and doesn't cause any noticiable performance degradations:

## Benchmarks (just in case)

Master:

Method | Mean | Error | StdDev | Allocated
-- | -- | -- | -- | --
FindPreResource | 283.12 μs | 4.266 μs | 3.990 μs | -
FindPostResource | 11.68 μs | 0.039 μs | 0.036 μs | -
FindNotExistingResource | 297.48 μs | 4.413 μs | 4.128 μs | -

Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
CreateButton | 27.16 μs | 0.103 μs | 0.091 μs | 1.8921 | 0.0610 | 34.85 KB

Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- |  -- | -- | -- | -- | -- | --
InitFluentTheme | 366,026.07 ns | 1,457.901 ns | 1,363.722 ns | 40.5273 | 19.5313 | 769608 B
InitSimpleTheme | 43,979.18 ns | 383.034 ns | 358.290 ns | 5.4932 | 1.4038 | 103400 B


This PR:

Method | Mean | Error | StdDev | Allocated
-- | -- | -- | -- | --
FindPreResource | 284.93 μs | 4.974 μs | 4.652 μs | -
FindPostResource | 12.17 μs | 0.053 μs | 0.049 μs | -
FindNotExistingResource | 310.56 μs | 6.187 μs | 8.045 μs | -

Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
CreateButton | 27.49 μs | 0.181 μs | 0.170 μs | 1.8921 | 0.0610 | 35.06 KB

Method | Mean | Error | StdDev | Gen0 | Gen1 | Allocated
-- | -- | -- | -- | -- | -- | --
InitFluentTheme | 361,861.45 ns | 1,878.032 ns | 1,756.712 ns | 40.5273 | 19.5313 | 769936 B
InitSimpleTheme |  44,886.64 ns | 769.670 ns | 719.950 ns | 5.4932 | 1.4038 | 103664 B

## Checklist

- [x] Added unit tests (if possible)?
- [x] Added XML documentation to any related classes?
- [-] Consider submitting a PR to https://github.com/AvaloniaUI/Documentation with user documentation

## Fixed issues

Fixes #10727